+++ backend/core/src/auth/dto/user-roles.dto.ts
@@ -1,7 +1,7 @@
 import { OmitType } from "@nestjs/swagger"
 import { UserRoles } from "../entities/user-roles.entity"
 
-export class UserRolesDto extends OmitType(UserRoles, ["isPartner", "isAdmin"] as const) {}
+export class UserRolesDto extends OmitType(UserRoles, ["user"] as const) {}
 
 export class UserRolesCreateDto extends OmitType(UserRolesDto, [] as const) {}