+++ backend/core/src/listings/dto/listing.dto.ts
@@ -33,6 +33,11 @@ import { AssetCreateDto, AssetDto, AssetUpdateDto } from "../../assets/dto/asset
 import { ApplicationMethodDto } from "../../application-methods/dto/application-method.dto"
 import { ListingEventCreateDto, ListingEventDto, ListingEventUpdateDto } from "./listing-event.dto"
 import { listingUrlSlug } from "../../shared/url-helper"
+import {
+  UnitsSummaryCreateDto,
+  UnitsSummaryDto,
+  UnitsSummaryUpdateDto,
+} from "../../units-summary/dto/units-summary.dto"
 
 export class ListingDto extends OmitType(Listing, [
   "applicationAddress",
@@ -50,6 +55,7 @@ export class ListingDto extends OmitType(Listing, [
   "property",
   "reservedCommunityType",
   "result",
+  "unitsSummary",
 ] as const) {
   @Expose()
   @IsDefined({ groups: [ValidationsGroupsEnum.default] })
@@ -319,6 +325,12 @@ export class ListingDto extends OmitType(Listing, [
     { toClassOnly: true }
   )
   urlSlug: string
+
+  @Expose()
+  @IsDefined({ groups: [ValidationsGroupsEnum.default] })
+  @ValidateNested({ groups: [ValidationsGroupsEnum.default] })
+  @Type(() => UnitsSummaryDto)
+  unitsSummary: UnitsSummaryDto[]
 }
 
 export class PaginatedListingDto extends PaginationFactory<ListingDto>(ListingDto) {}
@@ -354,6 +366,7 @@ export class ListingCreateDto extends OmitType(ListingDto, [
   "jurisdiction",
   "reservedCommunityType",
   "result",
+  "unitsSummary",
 ] as const) {
   @Expose()
   @IsDefined({ groups: [ValidationsGroupsEnum.default] })
@@ -513,6 +526,12 @@ export class ListingCreateDto extends OmitType(ListingDto, [
   @ValidateNested({ groups: [ValidationsGroupsEnum.default] })
   @Type(() => AssetCreateDto)
   result?: AssetCreateDto | null
+
+  @Expose()
+  @IsDefined({ groups: [ValidationsGroupsEnum.default] })
+  @ValidateNested({ groups: [ValidationsGroupsEnum.default] })
+  @Type(() => UnitsSummaryCreateDto)
+  unitsSummary: UnitsSummaryCreateDto[]
 }
 
 export class ListingUpdateDto extends OmitType(ListingDto, [
@@ -546,6 +565,7 @@ export class ListingUpdateDto extends OmitType(ListingDto, [
   "jurisdiction",
   "reservedCommunityType",
   "result",
+  "unitsSummary",
 ] as const) {
   @Expose()
   @IsOptional({ groups: [ValidationsGroupsEnum.default] })
@@ -722,6 +742,12 @@ export class ListingUpdateDto extends OmitType(ListingDto, [
   @ValidateNested({ groups: [ValidationsGroupsEnum.default] })
   @Type(() => AssetUpdateDto)
   result?: AssetUpdateDto
+
+  @Expose()
+  @IsDefined({ groups: [ValidationsGroupsEnum.default] })
+  @ValidateNested({ groups: [ValidationsGroupsEnum.default] })
+  @Type(() => UnitsSummaryUpdateDto)
+  unitsSummary: UnitsSummaryUpdateDto[]
 }
 
 // add other listing filter params here